hotfix mean malfunction

+++ simple_math.c
@@ -12,7 +12,7 @@
  * \return media 
  */
 float mean(int a, int b){
-    return (a + b)/2;
+    return (a + b)/(float)2;
 }
 
 int max(int a, int b){
